start to change system plugin structure to Joomla 4+: - add service provider - remove language prefixes in localisation files names - Replace main plugin file to src\Extension
