--- plg_system_jlcontentfieldsfilter/fields/index.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><title></title>
