--- plg_system_jlcontentfieldsfilter/language/en-GB/en-GB.plg_system_jlcontentfieldsfilter.sys.ini
@@ -1,4 +1,0 @@
-PLG_JLCONTENTFIELDSFILTER="System - JL Content Fields Filter"
-PLG_JLCONTENTFIELDSFILTER_XML_DESCRIPTION="Plugin for filtering Joomla content by fields"
-JOPTION_FROM_MODULE="From module"
-JOPTION_FROM_TEMPLATE="From template %s"
