+++ plg_system_jlcontentfieldsfilter/language/fr-FR/plg_system_jlcontentfieldsfilter.sys.ini
