+++ plg_system_jlcontentfieldsfilter/language/ru-RU/plg_system_jlcontentfieldsfilter.sys.ini
@@ -1,0 +1,4 @@
+PLG_JLCONTENTFIELDSFILTER="Система - JL Content Fields Filter"
+PLG_JLCONTENTFIELDSFILTER_XML_DESCRIPTION="Плагин для фильтрации материалов Joomla по дополнительным полям"
+JOPTION_FROM_MODULE="Из модуля"
+JOPTION_FROM_TEMPLATE="Из шаблона %s"
